Fix gzip magic byte check: cast to uint8_t (char is signed on x86)
ROOT CAUSE of gzip decompression never running:

The hex dump proved the body DID start with 0x1F 0x8B after chunked
decoding:
  [Browser] AFTER chunked: new_len=388, first 40 bytes hex:
  1F 8B 08 00 00 00 00 00 00 FF 65 52 CB 6E DC 30 ...

But the check 'body[0] == 0x1F && body[1] == 0x8B' was FALSE.

BUG: 'char' is signed by default on x86. body[1] is 0x8B, which as
a signed char is -117. When compared to the int literal 0x8B (139),
the signed char gets sign-extended to 0xFFFFFF8B (-117), which is
NOT equal to 0x8B (139). So the gzip magic byte check always failed
for any byte >= 0x80.

FIX: Cast body[n] to uint8_t before comparison:
  if (is_gzip && (uint8_t)body[0] == 0x1F && (uint8_t)body[1] == 0x8B)

This ensures the comparison is done as unsigned 0-255, matching the
int literals correctly.

This is the LAST bug preventing rendering. With this fix:
  1. Chunked decoding works (already fixed)
  2. Gzip magic byte check now passes
  3. gzip_inflate decompresses the body to actual HTML
  4. DOM parser finds <html>, <head>, <body>, <h1>Example Domain</h1>
  5. Page renders!

+++ usr/apps/browser.c
@@ -1090,7 +1090,11 @@ static void browser_load_page(const char* url) {
         }
     }
 
-    if (is_gzip && body[0] == 0x1F && body[1] == 0x8B) {
+    // Cast body[n] to uint8_t before comparison. `char` is signed on x86,
+    // so body[1] == 0x8B sign-extends to 0xFFFFFF8B (-117), which is NOT
+    // equal to 0x8B (139). This was causing the gzip magic byte check to
+    // fail even when the body correctly started with 0x1F 0x8B.
+    if (is_gzip && (uint8_t)body[0] == 0x1F && (uint8_t)body[1] == 0x8B) {
         // Decompress gzip body
         int body_len = total_read - (body - response);
         uint32_t decompressed_len = 0;
